HuBERT-ECG/code/requirements.txt
streamlit==1.30.0
numpy==1.24.3
matplotlib==3.7.1
Pillow==10.0.0
fpdf==1.7.2
torch==2.0.1
requests==2.31.0
neurokit2==0.2.4
PyWavelets==1.8.0 